requirements.txt
annotated-types==0.6.0
anyio==3.7.1
av==10.0.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==3.20.0
deepl==1.16.0
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
faster-whisper==0.9.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.10.0
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.16.1
orjson==3.9.10
packaging==23.2
protobuf==4.25.0
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.14.1
torch==2.1.0 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.24.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
webvtt-py==0.4.6
